pyproject.toml
[tool.poetry]
name = "self-corrective-rag-idp"
version = "0.1.0"
description = "RAG Implementation using FastAPI or Flask for contextual search of datasets on India Data Portal"
authors = ["Surya Kandukuri <sudham.surya@gmail.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
requests = ">=2.32.3,<3.0.0"
fastapi = ">=0.115.8,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
python-dotenv = ">=1.0.1,<2.0.0"
chromadb = "^0.6.3"
black = "^25.1.0"
langchain = "^0.3.21"
langchain-groq = "^0.3.1"
langchain-community = "^0.3.20"
langgraph = "^0.3.21"
langdetect = "^1.0.9"
langchain-experimental = "^0.3.4"
langchain-text-splitters = "^0.3.7"
nltk = "^3.9.1"
langchain-chroma = "^0.2.2"
ipykernel = "^6.29.5"
sentence-transformers = "^4.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
